requirements.txt
pypdf==3.17.4
python-dotenv==1.0.0
llama-index==0.9.27
discord-py-interactions==5.11.0